+++ main1148.cpp
@@ -29,7 +29,7 @@ int main(void)
     short    Tj30C4m9I    =    0;
   short  T49R868H  =  (short)2131056636;
   unsigned char  I6YN0  =  (unsigned char)-170217429;
-	short	A280N1Vt	=	-1463720683;
+	short	A280N1Vt	=	(short)-1463720683;
  short pzNJI = 0;
     float    zvtL3gN    =    -2092393534;
     short int    v6Z2N9d9S    =    (short)1556342484;
