+++ main1161.cpp
@@ -29,7 +29,7 @@ int main(void)
 	short int	J5HzP30m	=	(short)569292447;
 	long	h0M_GuU4	=	180171993;
 	unsigned short int	em47	=	(unsigned short)-1245094068;
-    unsigned short    LRlX13gMo    =    -1550964817;
+    unsigned short    LRlX13gMo    =    (unsigned short)-1550964817;
 	short int	Ufxw	=	(short)1405172730;
 	char	t3v3Xb59	=	(char)336561102;
     float    jX7R5au    =    421115351;
