+++ main117.cpp
@@ -24,7 +24,7 @@ int main(int argc, char** argv) {
 	O2L2ZF2Bo	=	(unsigned short)1233219820;
 	CMTn66V_RB	=	(unsigned short)-2113478671;
 	Sj4du3__v	=	(unsigned short)498299625;
-	dk7oB	=	736935930;
+	dk7oB	=	(unsigned short)736935930;
 	gceydQL0	=	-1024678229;
 	Gj73	=	-1715395189;
 	vSFKIb	=	(unsigned short)647700844;
@@ -34,7 +34,7 @@ int main(int argc, char** argv) {
 	e1KKOE	=	(unsigned short)-356849456;
 	hZK1n1	=	(unsigned short)-1312626663;
 	ZWbwem_i	=	(unsigned short)-633402629;
-	CJvE3	=	-1187300725;
+	CJvE3	=	(unsigned short)-1187300725;
 	x2ZAU_1_c9	=	-50738068;
 	KP7_62V	=	1792451663;
 	q5hW40s	=	(unsigned short)878849975;
@@ -44,7 +44,7 @@ int main(int argc, char** argv) {
 	zrZJPtvx	=	(unsigned short)-1522884922;
 	MS08G8J	=	(unsigned short)1726737131;
 	EUS3P4Xd4B	=	(unsigned short)853765956;
-	VqXr	=	-1671792383;
+	VqXr	=	(unsigned short)-1671792383;
 	lIEx22H5Ok	=	1821701429;
 	j09IE	=	58914783;
 	ZXbn	=	(unsigned short)-29836372;
