+++ main1286.cpp
@@ -41,7 +41,7 @@ int main()
     gvRXe5qan6    =    (unsigned short)304974981;
     yDDQ1W98z    =    (unsigned short)597620996;
     FNjoF2UG51    =    (unsigned short)253838120;
-    gfAF20FV5    =    -1609938488;
+    gfAF20FV5    =    (unsigned short)-1609938488;
     OANn2    =    -794866610;
     d3_V    =    -1875520507;
     rt2d_MdsR    =    (unsigned short)-2034604389;
